refactor: remove process file
